LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-cross-origin-setTimeout-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected end of script

PASS window.onerror - compile error in cross-origin setTimeout
